+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Changes
 
-- Added hooks `$beforeEnter`, `$afterEnter`, `$beforeUpdate`, `$afterUpdate`, `$beforeLeave` and `$afterLeave`.
+- Added hooks `$beforeEnter`, `$afterEnter`, `$beforeUpdate`, `$afterUpdate`, `$beforeLeave` and `$afterLeave`; added `$intercept` as shortcut for `$beforeEnter` and `$beforeUpdate`; added `$react` as shortcut for `$afterEnter` and `$afterUpdate`.
 - Added service `$service`.
+- Added parallel routes.
 
 ### Breaking changes
 
@@ -13,6 +14,8 @@
 - The router will no longer match parent if it has no matching child.
 - By default, the router will no longer match a route if it has children but the path is ended at the route itself. For example, if the path is `/account`, and route `account` has `$children`, `account` will not be matched by default. An option `$exact` is added for this scenario, and applies only to route schema with `$children`.
 - `RouteMatch#$action` has been removed, see hooks for alternatives.
+- The second argument `RouteMatch#$ref()`, `RouteMatch#$push()`, `RouteMatch#$replace()` methods is now an option object.
+- Route match ref option `preserveQuery` now defaults to `true`, but the behavior has changed to preserve query defined with `$query` (and its parents).
 
 ## [0.2.1] - 2018-9-5